Detach CrashView from CrashHandler
