+++ crash-handler/src/test/java/org/odk/collect/crashhandler/CrashHandlerTest.kt
@@ -44,125 +44,6 @@ class CrashHandlerTest {
     }
 
     @Test
-    fun getCrashView_whenThereAreNoCrashesRegistered_returnsNull() {
-        assertThat(createCrashHandler().getCrashView(context), equalTo(null))
-    }
-
-    @Test
-    fun getCrashView_whenThereHasBeenACrashRegistered_returnsViewWithCrashDetails() {
-        createCrashHandler().registerCrash(context, RuntimeException("crash!"))
-
-        val view = createCrashHandler().getCrashView(context)
-        assertThat(
-            view!!.findViewById<TextView>(R.id.title).text,
-            equalTo(context.getString(org.odk.collect.strings.R.string.crash_last_run))
-        )
-        assertThat(view.findViewById<TextView>(R.id.message).text, equalTo("crash!"))
-    }
-
-    @Test
-    fun getCrashView_whenThereHasBeenACrashRegistered_andTheOkButtonIsClickedOnTheView_runsTheOnErrorDismissedListener() {
-        createCrashHandler().registerCrash(context, RuntimeException())
-
-        val onErrorDismissed = mock<Runnable>()
-        val view = createCrashHandler().getCrashView(context, onErrorDismissed)
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        verify(onErrorDismissed).run()
-    }
-
-    @Test
-    fun getCrashView_whenThereHasBeenACrashRegistered_andTheOkButtonIsClickedOnTheView_doesNotRunProcessKiller() {
-        createCrashHandler().registerCrash(context, RuntimeException())
-
-        val onErrorDismissed = mock<Runnable>()
-        val view = createCrashHandler().getCrashView(context, onErrorDismissed)
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        verify(processKiller, never()).run()
-    }
-
-    @Test
-    fun getCrashView_whenThereHasBeenACrashRegistered_andTheOkButtonIsClickedOnTheView_clearsCrash() {
-        createCrashHandler().registerCrash(context, RuntimeException())
-
-        // Check crash view is returned for multiple instances (for rotates etc)
-        assertThat(createCrashHandler().getCrashView(context), notNullValue())
-
-        val onErrorDismissed = mock<Runnable>()
-        val view = createCrashHandler().getCrashView(context, onErrorDismissed)
-
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        assertThat(createCrashHandler().getCrashView(context), equalTo(null))
-    }
-
-    @Test
-    fun getCrashView_whenCheckingConditionsFails_returnsViewWithCrashDetails() {
-        val crashHandler = createCrashHandler()
-
-        crashHandler.launchApp({
-            throw RuntimeException("blah")
-        })
-
-        val view = crashHandler.getCrashView(context)
-        assertThat(
-            view!!.findViewById<TextView>(R.id.title).text,
-            equalTo(context.getString(org.odk.collect.strings.R.string.cant_start_app))
-        )
-        assertThat(view.findViewById<TextView>(R.id.message).text, equalTo("blah"))
-    }
-
-    @Test
-    fun getCrashView_whenCheckingConditionsFails_andTheOkButtonIsClickedOnTheView_killsTheProcess() {
-        val crashHandler = createCrashHandler()
-
-        crashHandler.launchApp({
-            throw RuntimeException("blah")
-        })
-
-        val view = crashHandler.getCrashView(context)
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        verify(processKiller).run()
-    }
-
-    @Test
-    fun getCrashView_whenCheckingConditionsFails_andTheOkButtonIsClickedOnTheView_doesNotRunErrorDismissedListener() {
-        val crashHandler = createCrashHandler()
-
-        crashHandler.launchApp({
-            throw RuntimeException("blah")
-        })
-
-        val onErrorDismissed = mock<Runnable>()
-        val view = crashHandler.getCrashView(context, onErrorDismissed)
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        verify(onErrorDismissed, never()).run()
-    }
-
-    @Test
-    fun getCrashView_whenCheckConditionFailedInDifferentInstance_returnsNull() {
-        val crashHandler = createCrashHandler()
-        crashHandler.launchApp({
-            throw RuntimeException("blah")
-        })
-
-        val otherCrashHandler = createCrashHandler()
-        assertThat(otherCrashHandler.getCrashView(context), equalTo(null))
-    }
-
-    @Test
-    fun getCrashView_whenThereHasBeenACrashRegistered_andCheckConditionFailed_andTheOkButtonIsClickedOnTheView_runsProcessKiller() {
-        createCrashHandler().registerCrash(context, RuntimeException())
-
-        val crashHandler = createCrashHandler()
-        crashHandler.launchApp({
-            throw RuntimeException("blah")
-        })
-
-        val view = crashHandler.getCrashView(context)
-        view!!.findViewById<View>(R.id.ok_button).performClick()
-        verify(processKiller).run()
-    }
-
-    @Test
     fun launchApp_whenConditionsFail_runsOnSuccess() {
         val crashHandler = createCrashHandler()
